+++ stroom-ui/src/components/Account/TokenManager/TokenManager.tsx
@@ -97,6 +97,7 @@ const TokenManager: FunctionComponent<{
           },
           quickFilterProps,
           pagerProps,
+          keyExtractor: (token) => token.id.toString(),
         }}
         onClose={onClose}
       />